1.0 add correct what vsevolod sow

+++ src/main/java/ua/bank/moneyguard/dtos/mainDTOs/CardDataDTO.java
@@ -13,6 +13,6 @@ public class CardDataDTO {
     private String cardNumber;
     private LocalDate validity;
     private Integer cvv;
-    private Long amountOfMoney;
+    private Double amountOfMoney;
     private String ownerName;
 }
